Refactor dependencies to facilitate reuse
This allows us to export the singletons and classes to the other modules.

See https://code.visualstudio.com/docs/extensionAPI/vscode-api#_extensions
