+++ packages/salesforcedx-utils-vscode/src/types/index.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2017, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+// Precondition checking
+////////////////////////
+export interface PreconditionChecker {
+  check(): boolean;
+}
+
+export interface PostconditionChecker<T> {
+  check(
+    inputs: ContinueResponse<T> | CancelResponse
+  ): Promise<ContinueResponse<T> | CancelResponse>;
+}
+
+// Input gathering
+//////////////////
+export interface ContinueResponse<T> {
+  type: 'CONTINUE';
+  data: T;
+}
+
+export interface CancelResponse {
+  type: 'CANCEL';
+}
+
+export interface ParametersGatherer<T> {
+  gather(): Promise<CancelResponse | ContinueResponse<T>>;
+}
+
+// Selection
+////////////
+
+export type DirFileNameSelection = {
+  fileName: string;
+  outputdir: string;
+};
