+++ packages/salesforcedx-vscode-core/src/channels/index.ts
@@ -6,6 +6,6 @@
  */
 
 import { ChannelService } from './channelService';
-
+export { ChannelService } from './channelService';
 export const channelService = ChannelService.getInstance();
 export { DEFAULT_SFDX_CHANNEL } from './channelService';
