+++ packages/salesforcedx-vscode-core/src/commands/forceGenerateFauxClasses.ts
@@ -12,10 +12,10 @@ import {
   LocalCommandExecution,
   SfdxCommandBuilder
 } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import * as vscode from 'vscode';
 import { nls } from '../messages';
 import {
-  ContinueResponse,
   EmptyParametersGatherer,
   SfdxCommandlet,
   SfdxCommandletExecutor,
