+++ packages/salesforcedx-vscode-core/src/commands/forceLightningEventCreate.ts
@@ -10,6 +10,10 @@ import {
   Command,
   SfdxCommandBuilder
 } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import {
+  ContinueResponse,
+  DirFileNameSelection
+} from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import * as path from 'path';
 import { Observable } from 'rxjs/Observable';
 import * as vscode from 'vscode';
@@ -19,8 +23,6 @@ import { notificationService } from '../notifications';
 import { CancellableStatusBar, taskViewService } from '../statuses';
 import {
   CompositeParametersGatherer,
-  ContinueResponse,
-  DirFileNameSelection,
   LightningFilePathExistsChecker,
   SelectFileName,
   SelectStrictDirPath,
