+++ packages/salesforcedx-vscode-core/src/commands/forceProjectCreate.ts
@@ -10,6 +10,12 @@ import {
   Command,
   SfdxCommandBuilder
 } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import {
+  CancelResponse,
+  ContinueResponse,
+  ParametersGatherer,
+  PostconditionChecker
+} from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import * as fs from 'fs';
 import * as path from 'path';
 import { Observable } from 'rxjs/Observable';
@@ -19,12 +25,8 @@ import { nls } from '../messages';
 import { notificationService } from '../notifications';
 import { CancellableStatusBar, taskViewService } from '../statuses';
 import {
-  CancelResponse,
   CompositeParametersGatherer,
-  ContinueResponse,
   EmptyPreChecker,
-  ParametersGatherer,
-  PostconditionChecker,
   SfdxCommandlet,
   SfdxCommandletExecutor
 } from './commands';
