+++ packages/salesforcedx-vscode-core/src/notifications/index.ts
@@ -6,5 +6,6 @@
  */
 
 import { NotificationService } from './notificationService';
+export { NotificationService } from './notificationService';
 
 export const notificationService = NotificationService.getInstance();
