+++ packages/salesforcedx-vscode-core/src/statuses/index.ts
@@ -11,5 +11,6 @@ export {
   CancellableStatusBar
 } from './statusBar';
 import { TaskViewService } from './taskView';
+export { TaskViewService } from './taskView';
 export const taskViewService = TaskViewService.getInstance();
 export { Task } from './taskView';
