+++ packages/salesforcedx-vscode-core/test/commands/commands.test.ts
@@ -5,21 +5,23 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import {
+  CancelResponse,
+  ContinueResponse,
+  DirFileNameSelection,
+  ParametersGatherer
+} from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import { expect } from 'chai';
 import * as path from 'path';
 import * as sinon from 'sinon';
 import * as vscode from 'vscode';
 import {
-  CancelResponse,
   CommandletExecutor,
   CompositeParametersGatherer,
-  ContinueResponse,
-  DirFileNameSelection,
   EmptyParametersGatherer,
   EmptyPostChecker,
   FilePathExistsChecker,
   LightningFilePathExistsChecker,
-  ParametersGatherer,
   SelectPrioritizedDirPath,
   SelectStrictDirPath,
   SfdxCommandlet
@@ -53,7 +55,7 @@ describe('Command Utilities', () => {
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             throw new Error('This should not be called');
           }
         }(),
@@ -77,7 +79,7 @@ describe('Command Utilities', () => {
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             return { type: 'CANCEL' };
           }
         }(),
@@ -102,7 +104,7 @@ describe('Command Utilities', () => {
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             return { type: 'CONTINUE', data: {} };
           }
         }(),
@@ -125,14 +127,14 @@ describe('Command Utilities', () => {
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             return { type: 'CONTINUE', data: {} };
           }
         }(),
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             return { type: 'CONTINUE', data: {} };
           }
         }()
@@ -147,14 +149,14 @@ describe('Command Utilities', () => {
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             return { type: 'CANCEL' };
           }
         }(),
         new class implements ParametersGatherer<{}> {
           public async gather(): Promise<
             CancelResponse | ContinueResponse<{}>
-            > {
+          > {
             throw new Error('This should not be called');
           }
         }()
@@ -175,7 +177,7 @@ describe('Command Utilities', () => {
           new class implements ParametersGatherer<{}> {
             public async gather(): Promise<
               CancelResponse | ContinueResponse<{}>
-              > {
+            > {
               return { type: 'CONTINUE', data: {} };
             }
           }()
@@ -203,7 +205,7 @@ describe('Command Utilities', () => {
           new class implements ParametersGatherer<{}> {
             public async gather(): Promise<
               CancelResponse | ContinueResponse<{}>
-              > {
+            > {
               return { type: 'CANCEL' };
             }
           }()
