+++ packages/salesforcedx-vscode-core/test/commands/forceDebuggerStop.test.ts
@@ -5,10 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 import {
-  ContinueResponse,
   SfdxCommandlet,
   SfdxWorkspaceChecker
 } from '../../src/commands/commands';
